% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStats_MLET.R
\name{ClusterStats_MLET}
\alias{ClusterStats_MLET}
\title{Generate unique random permutations}
\usage{
ClusterStats_MLET(data, paired = T, detailed = F, threshold_t = NA)
}
\description{
Generate unique random permutations
}
