% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTimeBinData_MLET.R
\name{CreateTimeBinData_MLET}
\alias{CreateTimeBinData_MLET}
\title{Collapse time points in specified time bins}
\usage{
CreateTimeBinData_MLET(
  data,
  groupingColumns = NULL,
  timeBinWidth = 250,
  timeMax = 3000,
  FixatedOn = "FixatedOn",
  timepoint = "timepoint",
  AOIs = NULL,
  timeForward = T,
  aggregateFun = mean
)
}
\description{
Collapse time points in specified time bins
}
