% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataForMATLAB_MLET.R
\name{ExportDataForMATLAB_MLET}
\alias{ExportDataForMATLAB_MLET}
\title{Export to CSV file for MATLAB GUI}
\usage{
ExportDataForMATLAB_MLET(
  data,
  GazeX = "GazeX_Relative",
  GazeY = "GazeY_Relative",
  AOINames = NULL,
  fileName = "ETDataforMATLAB.csv",
  path = getwd()
)
}
\description{
Export to CSV file for MATLAB GUI
}
