% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterClustersTimepoints.R
\name{FilterClustersTimepoints}
\alias{FilterClustersTimepoints}
\title{Generate unique random permutations}
\usage{
FilterClustersTimepoints(data, clusterInf = clusterInf)
}
\description{
Generate unique random permutations
}
