% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationTest_MLET.R
\name{PermutationTest_MLET}
\alias{PermutationTest_MLET}
\title{Permutation Test General Routine}
\usage{
PermutationTest_MLET(
  data,
  samples = 2000,
  paired = T,
  permuteTrialsWithinSubject = F,
  threshold_t = NA
)
}
\description{
Permutation Test General Routine
}
