% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTemporalGazeTrends_MLET.R
\name{PlotTemporalGazeTrends_MLET}
\alias{PlotTemporalGazeTrends_MLET}
\title{Plot temporal gaze trends}
\usage{
PlotTemporalGazeTrends_MLET(
  resultList,
  showDataPointNumbers = T,
  gazePropRibbonAlpha = 0.1,
  clusterFillColor = "#CC9933",
  clusterFillAlpha = 0.5,
  testNameOrder = NULL,
  conditionOrder = NULL,
  onlySignificantClusters = T,
  clusterData = NULL
)
}
\arguments{
\item{resultList}{can be a dataframe of the data that is already prepared by PrepareMLETData_MLET or a list that is the result of PermutationTest_MLET}
}
\description{
Plot temporal gaze trends
}
