% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareMLETData_MLET.R
\name{PrepareMLETData_MLET}
\alias{PrepareMLETData_MLET}
\title{Prepare dataa columns}
\usage{
PrepareMLETData_MLET(
  data,
  ID,
  trial,
  timepoint,
  condition,
  AOI,
  conditionLevels = NA,
  targetAOI = NA
)
}
\description{
Prepare dataa columns
}
