% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessRawData_MLET.R
\name{PreprocessRawData_MLET}
\alias{PreprocessRawData_MLET}
\title{Pre-Process data to extract AOIs}
\usage{
PreprocessRawData_MLET(
  data,
  GazeX = "GazeX_Relative",
  GazeY = "GazeY_Relative",
  AOINames = NULL,
  fileName = "ETDataforMATLAB.csv",
  path = getwd()
)
}
\description{
Pre-Process data to extract AOIs
}
