% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialLevelPermutationTestWithin_MLET.R
\name{TrialLevelPermutationTestWithin_MLET}
\alias{TrialLevelPermutationTestWithin_MLET}
\title{Generate unique random permutations}
\usage{
TrialLevelPermutationTestWithin_MLET(
  data,
  samples = 2000,
  paired = T,
  threshold_t = NA
)
}
\description{
Generate unique random permutations
}
