% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataForMATLAB_MLET.R
\name{ExportDataForMATLAB_MLET}
\alias{ExportDataForMATLAB_MLET}
\title{Prepare data columns}
\usage{
ExportDataForMATLAB_MLET(
  data,
  fileName = "ETDataforMATLAB.csv",
  path = getwd()
)
}
\description{
Prepare data columns
}
